renamed home to result as home dose not mak any sense, also wrote some todos
